Add external network E2E to flake finder
Include new external network deployments and tests in the periodic tests
against devel meant to find flaky failures.

Make minor inline-docs clarifications to both external E2E jobs.

I'd like to be able to just `include` these params in the existing E2E
matrix to create new jobs, but need to be able to skip the external
tests for those jobs but not these. If the external tests could skip
themselves when they are not relevant, this duplication could be
avoided. That might not be worth the added fragility/complexity though.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>