Refactor NAT discovery unit tests to use the public Interface
This makes them more "whitebox" and moves the files to the
"natdiscovery_test" package to avoid leakage of test code into
the production package. This should also make a bit easier to
implement IPv6.

Two of the test files, request_handle_internal_test.go and
listener_internal_test.go, were not moved to the test package as
they test specific internal functions which would entail making
too much of the production code public.

A new function, NewWithConfig, was added that allows the unit tests to
construct the natDiscovery instance with hooks to mock behavior.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>